requirements.txt
elasticsearch==7.5.1
prometheus-client==0.7.1
pyaml==19.12.0
PyYAML==5.3
setproctitle==1.1.10
urllib3==1.25.8
